refactor: Split MainPage into server and client components
- Extracted all client-side logic and state management from MainPage
  into a new MainPageClient component with 'use client' directive.
- Updated MainPage to be a server component that uses Suspense to
  asynchronously load MainPageClient with a loading fallback.
- This separation resolves build errors related to client-only hooks
  and improves code organization.
- Added a simple Loading component as Suspense fallback.
- Introduced a GitHub Actions workflow for CI to build and test the
  Next.js project across multiple Node.js versions.
