+++ .github/workflows/npm_build.yaml
@@ -1,0 +1,40 @@
+name: plateeRAG_CI
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  build-and-test:
+    runs-on: ubuntu-latest
+
+    strategy:
+      matrix:
+        node-version: [18.x, 20.x, 22.x, 24.x]
+
+    steps:
+      - name: Checkout source code
+        uses: actions/checkout@v4
+
+      - name: Use Node.js ${{ matrix.node-version }}
+        uses: actions/setup-node@v4
+        with:
+          node-version: ${{ matrix.node-version }}
+          cache: 'npm'
+
+      - name: Install dependencies
+        run: npm ci
+
+      - name: Build Next.js project
+        run: npm run build
+
+    #   - name: Run tests
+    #     run: npm test
+
+      - name: Start Next.js server and check status
+        run: |
+          npm run start & # 백그라운드에서 서버 실행
+          sleep 5 # 서버가 시작될 때까지 잠시 대기
+          curl -f http://localhost:3000 # localhost:3000에 접속하여 HTTP 200 응답 확인
